requirements.txt
PyJWT~=1.0
